selectFieldCommit
