+++ src/main/java/com/example/team_project/camp/CampService.java
@@ -1,7 +1,6 @@
 package com.example.team_project.camp;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
@@ -9,10 +8,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.example.team_project._core.erroes.exception.Exception404;
 import com.example.team_project.camp._dto.CampReqDTO;
-import com.example.team_project.camp._dto.CampReqDTO.CampBookmarkDTO;
 import com.example.team_project.camp._dto.CampRespDTO;
 import com.example.team_project.camp._dto.CampRespDTO.CampDetailDTO;
-import com.example.team_project.camp._dto.CampRespDTO.CampListDTO;
 import com.example.team_project.camp.camp_bookmark.CampBookmark;
 import com.example.team_project.camp.camp_bookmark.CampBookmarkJPARepository;
 import com.example.team_project.camp.camp_image.CampImage;
@@ -22,9 +19,6 @@ import com.example.team_project.camp.camp_review.CampReview;
 import com.example.team_project.camp.camp_review.CampReviewJPARepository;
 import com.example.team_project.camp_field.CampField;
 import com.example.team_project.camp_field.CampFieldJPARepository;
-import com.example.team_project.camp_field._dto.CampFieldReqDTO;
-import com.example.team_project.camp_field._dto.CampFieldRespDTO;
-import com.example.team_project.camp_field._dto.CampFieldRespDTO.CampFieldListDTO;
 import com.example.team_project.user.User;
 import com.example.team_project.user.UserJPARepository;
 
@@ -121,16 +115,5 @@ public class CampService {
         return new CampRespDTO.MyCampListDTO(campReviews, requestDTO.getYear());
     }
 
-	
-	//캠프장 아이디를 받아 캠프 구역 목록 조회 + 캠프장 지도 + 상세정보 조회
-	public CampRespDTO.CampFieldListDTO  campFieldList(CampReqDTO.CampFieldListDTO requestDTO) {
-		// 캠프장 정보 조회
-		Camp camp = campJPARepository.findById(requestDTO.getCampId()).orElseThrow(() ->
-				new Exception404("해당 캠프장이 존재하지 않습니다."));
-		// 캠프 구역 목록 조회
-		List<CampField> campfields = campFieldJPARepository.findAllByCampId(requestDTO.getCampId());
-		if(campfields == null)throw new Exception404("잘못된 캠프장 명입니다.");
-		return new CampRespDTO.CampFieldListDTO(campfields, camp, requestDTO);
-	}
 
 }
