+++ src/main/java/com/example/team_project/camp/_dto/CampReqDTO.java
@@ -16,11 +16,6 @@ public class CampReqDTO {
 	}
 
 	
-	@Data
-	public static class CampFieldListDTO{
-		private Integer campId; // 캠프장 아이디
-		private String checkInDate; //체크인 날짜
-		private String checkOutDate; //체크아웃 날짜
-	}
+	
 	
 }
